+++ sdk/core/Azure.Core.Experimental/src/DynamicData/MutableJsonElement.WriteTo.cs
