+++ sdk/core/Azure.Core.Experimental/tests/public/Azure.Core.Experimental.Tests.Public.csproj
@@ -11,7 +11,7 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="$(AzureCoreTestFramework)" />
-    <ProjectReference Include="..\..\src\Azure.Core.csproj" />
+    <ProjectReference Include="..\..\src\Azure.Core.Experimental.csproj" />
   </ItemGroup>
   
   <!-- Shared source from Azure.Core -->
