+++ sdk/core/Azure.Core.Experimental/tests/public/JsonDataPublicMutableTests.cs
