+++ sdk/core/Azure.Core.Experimental/tests/public/JsonDataTestHelpers.cs
