+++ sdk/core/Azure.Core/Azure.Core.sln
@@ -51,7 +51,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.Core.Expressions.Data
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.Core.Expressions.DataFactory.Tests", "..\Azure.Core.Expressions.DataFactory\tests\Azure.Core.Expressions.DataFactory.Tests.csproj", "{E1E88CD0-609C-4B8C-B723-DB004B90E235}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.Core.Tests.Public", "tests\public\Azure.Core.Tests.Public.csproj", "{132374AD-FF07-40B0-887E-A73D45C91537}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.Core.Experimental.Tests.Public", "..\Azure.Core.Experimental\tests\public\Azure.Core.Experimental.Tests.Public.csproj", "{00955D8A-B387-474B-9FA9-449A1013990F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -127,10 +127,10 @@ Global
 		{E1E88CD0-609C-4B8C-B723-DB004B90E235}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E1E88CD0-609C-4B8C-B723-DB004B90E235}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E1E88CD0-609C-4B8C-B723-DB004B90E235}.Release|Any CPU.Build.0 = Release|Any CPU
-		{132374AD-FF07-40B0-887E-A73D45C91537}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{132374AD-FF07-40B0-887E-A73D45C91537}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{132374AD-FF07-40B0-887E-A73D45C91537}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{132374AD-FF07-40B0-887E-A73D45C91537}.Release|Any CPU.Build.0 = Release|Any CPU
+		{00955D8A-B387-474B-9FA9-449A1013990F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{00955D8A-B387-474B-9FA9-449A1013990F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{00955D8A-B387-474B-9FA9-449A1013990F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{00955D8A-B387-474B-9FA9-449A1013990F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
