+++ sdk/core/Azure.Core/src/RequestContent.cs
@@ -8,7 +8,6 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Azure.Core.Buffers;
-using Azure.Core.Dynamic;
 using Azure.Core.Serialization;
 
 namespace Azure.Core
@@ -73,13 +72,6 @@ namespace Azure.Core
         public static RequestContent Create(BinaryData content) => new MemoryContent(content.ToMemory());
 
         /// <summary>
-        /// Creates an instance of <see cref="RequestContent"/> that wraps a <see cref="DynamicData"/>.
-        /// </summary>
-        /// <param name="content">The <see cref="DynamicData"/> to use.</param>
-        /// <returns>An instance of <see cref="RequestContent"/> that wraps a <see cref="DynamicData"/>.</returns>
-        public static RequestContent Create(DynamicData content) => new DynamicDataContent(content);
-
-        /// <summary>
         /// Creates an instance of <see cref="RequestContent"/> that wraps a serialized version of an object.
         /// </summary>
         /// <param name="serializable">The <see cref="object"/> to serialize.</param>
@@ -98,12 +90,6 @@ namespace Azure.Core
         /// </summary>
         /// <param name="content">The <see cref="BinaryData"/> to use.</param>
         public static implicit operator RequestContent(BinaryData content) => Create(content);
-
-        /// <summary>
-        /// Creates a RequestContent that wraps a <see cref="DynamicData"/>.
-        /// </summary>
-        /// <param name="content">The <see cref="DynamicData"/> to use.</param>
-        public static implicit operator RequestContent(DynamicData content) => Create(content);
 
         /// <summary>
         /// Writes contents of this object to an instance of <see cref="Stream"/>.
@@ -158,7 +144,8 @@ namespace Azure.Core
                     {
                         CancellationHelper.ThrowIfCancellationRequested(cancellationToken);
                         var read = _stream.Read(buffer, 0, buffer.Length);
-                        if (read == 0) { break; }
+                        if (read == 0)
+                        { break; }
                         CancellationHelper.ThrowIfCancellationRequested(cancellationToken);
                         stream.Write(buffer, 0, read);
                     }
@@ -280,35 +267,6 @@ namespace Azure.Core
             public override async Task WriteToAsync(Stream stream, CancellationToken cancellation)
             {
                 await stream.WriteAsync(_bytes, cancellation).ConfigureAwait(false);
-            }
-        }
-
-        private sealed class DynamicDataContent : RequestContent
-        {
-            private readonly DynamicData _data;
-
-            public DynamicDataContent(DynamicData data) => _data = data;
-
-            public override void Dispose()
-            {
-                _data.Dispose();
-            }
-
-            public override void WriteTo(Stream stream, CancellationToken cancellation)
-            {
-                _data.WriteTo(stream);
-            }
-
-            public override bool TryComputeLength(out long length)
-            {
-                length = default;
-                return false;
-            }
-
-            public override Task WriteToAsync(Stream stream, CancellationToken cancellation)
-            {
-                _data.WriteTo(stream);
-                return Task.CompletedTask;
             }
         }
     }
