+++ sdk/core/Azure.Core/src/Serialization/AzureCoreExtensions.cs
@@ -6,8 +6,6 @@ using System.Collections.Generic;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
-using Azure.Core.Dynamic;
-using Azure.Core.Json;
 using Azure.Core.Serialization;
 
 namespace Azure
@@ -58,15 +56,6 @@ namespace Azure
         {
             JsonElement element = data.ToObjectFromJson<JsonElement>();
             return element.GetObject();
-        }
-
-        /// <summary>
-        /// Return the content of the BinaryData as a dynamic type.
-        /// </summary>
-        public static dynamic ToDynamicFromJson(this BinaryData utf8Json)
-        {
-            MutableJsonDocument mdoc = MutableJsonDocument.Parse(utf8Json, DynamicData.DefaultSerializerOptions);
-            return new DynamicData(mdoc.RootElement);
         }
 
         private static object? GetObject(in this JsonElement element)
